+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/experimental/clink/debits/DebitEvent.kt
@@ -21,7 +21,8 @@
 package com.vitorpamplona.quartz.experimental.clink.debits
 
 import androidx.compose.runtime.Immutable
-import com.vitorpamplona.quartz.experimental.clink.Clink
+import com.vitorpamplona.quartz.experimental.clink.tags.ClinkVersionTag
+import com.vitorpamplona.quartz.experimental.clink.tags.clinkVersion
 import com.vitorpamplona.quartz.nip01Core.core.Event
 import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.core.OptimizedJsonMapper
@@ -58,7 +59,7 @@ class DebitEvent(
 
     fun isResponse() = requestId() != null
 
-    fun version() = tags.firstOrNull { it.size > 1 && it[0] == Clink.VERSION_TAG_NAME }?.get(1)
+    fun version() = tags.firstNotNullOfOrNull(ClinkVersionTag::parse)
 
     private fun talkingWith(oneSideHex: String): HexKey = if (pubKey == oneSideHex) recipientPubKey() ?: pubKey else pubKey
 
@@ -88,7 +89,7 @@ class DebitEvent(
             return signer.sign(
                 eventTemplate(KIND, encrypted, createdAt) {
                     pTag(servicePubKey, null)
-                    add(Clink.versionTag())
+                    clinkVersion()
                     alt(ALT)
                 },
             )
@@ -107,7 +108,7 @@ class DebitEvent(
                 eventTemplate(KIND, encrypted, createdAt) {
                     pTag(requestorPubKey, null)
                     add(ETag.assemble(requestEvent.id, null, null))
-                    add(Clink.versionTag())
+                    clinkVersion()
                     alt(ALT)
                 },
             )
